refactor(edge/devices): break filter and loading hook
